bgzf/virtual_position: Change the return value of the uncompressed position to u16
The uncompressed position cannot be larger than 2^16 - 1.

+++ noodles-bgzf/src/reader.rs
@@ -112,7 +112,7 @@ where
 
         self.block
             .data_mut()
-            .seek(SeekFrom::Start(uncompressed_offset))?;
+            .seek(SeekFrom::Start(u64::from(uncompressed_offset)))?;
 
         Ok(pos)
     }
